Extract version file, bump to 0.6.3.RC1 (read on)
Even though master version was 0.6.0 and the most recent release is 0.6.0
(not a very good practice to keep master version the same until release),
there are two tags in the repo: 0.6.1 and 0.6.2. I am not sure what to
do about this. Both tags date back to Dec 2009 or so.

So lets call it 0.6.3.RC1 and be done with it.