Fix the :timeout option to subscribe
send_command catches Qrack::ClientTimeout, in order to support
@read_write_timeout.  But ClientTimeout can be raised for another reason:
To support the :timeout option in Qrack::Subscription#start.  That
gets mishandled, so using the :timeout option can result in the connection
getting closed.

This separates out the two kinds of timeout.  :socket_timeout continues to
use ClientTimeout.  But the :timeout option now uses FrameTimeout
(as it indicates a timeout in next_frame).