Add a quick option for pass-through of ssl_context.
I'm messing around with some SSL options that are not currently
supported by Bunny, like client certificate authentication. The simplest
way I found to get started is this commit, which lets me create an
OpenSSL context object ahead of time, then pass it through to Bunny's
socket calls.

In the long run, this might be the best play. The alternative is
supporting all of the fiddly little ways people use SSL and, especially
fiddly, SSL authentication.