Parse a few more AMQP datatypes
Consuming from a RabbitMQ broker that is using the federation plugin
adds an x-received-from: header to messages that currently consists of
an array containing a table that contains a bunch of fields about the
origin broker.

Add support to parse all of the necessary datatypes so these messages
don't generate exceptions.