Alternate thread-safe cancellation of subscribe loops.
This implementation allows you to cancel subscribe loops by raising an
exception. This is an alternative to the IO based implementation
currently on master which I didn't like because of its complexity.

Example:
  subscriber = Thread.new
    queue.subscribe do |message|
      process(message)
    end
  end

  ...
  subscriber.raise Bunny::ForceUnsubscribe

The exception will make it to the block given to subscribe, which can
silence the exception or perform custom logic.

Example:
  subscriber = Thread.new
    queue.subscribe do |message|
      begin
        process(message)
      rescue Bunny::ForceUnsubscribe
        graceful_cleanup(message)
      end
    end
  end