Changed Client#next_frame's timeout and cancellator behavior
It was possible for the timeout to be hit while in the middle of
parsing, causing the buffer to be in a degraded state. This is also the
case when using cancellator, although in practice this is much harder to
do.

This change makes it so that the timeout or cancellator object only
cancels a subscription if it is in the waiting state, not when parsing.
This also has the added benefit of making the cancellator logic much
simplier, and both checks are done with a single method call.

Note that this implementation currently throws an exception if the
unsubscribe was unsucessful, which can happen if the timeout was hit
while the message was being delivered (network latency). This is far
better than indefinitely hanging however...