Initial migration to internal event loop for network activity and protocol method dispatch
Blocking operations remain blocking but no longer perform network operations
on the same thread. Instead, it is handled in a separate thread in a loop.
This is similar to how RabbitMQ Java client works internally and provides
a number of important benefits:

 * Ability to support server-initiated methods (basic.return, publisher confirms, cancellation notification, etc)

 * Avoids blocking publishing from within consumers

 * Better resource utilization on multicore systems (on JRuby and Rubinius)

 * Better separation of networking-related code from command dispatching
   and internal state changes in channels and connections

This commit only covers a few queue.* operations but the rest will look very
similarly.