Fix heartbeats, allow :heartbeat => 0 to disable heartbeats again
Previously with the value of 0 we used the RabbitMQ heartbeat value.
With 3.0, it is 600 seconds. That is a fine default, however, historically
amqp gem and Bunny used the value of 0 to disable heartbeats. This commit
makes that possible again. In addition, to keep the (very reasonable)
broker default in place, we change default heartbeat value to 600, which
means w/o :heartbeat option specified, heartbeats will be used. This
helps with firewalls closing idle connections and similar matters.

On the heartbeat sender implementation side, we should send heartbeat
frames roughly at 1/2 the heartbeat interval, not the full interval,
to avoid connection closure.