If connection is down, make Bunny::Queues#pop return empty responses
To support the same edge case with Bunny::Queue#pop being called
in a tight loop when network connection goes down.

It is better to return something immediately rather than keep
blocking the thread we have no easy way to unblock with the
current "blocking pop" approach that replaces not-entirely-correct
wait/notify implementation.