Perform automatic network recovery with a delay
Otherwise there is a race between the network activity thread that
tries to recover and the "main" connection thread. If the connection
is forced to shut down via the Management UI, we want to fail and
not create endlessly reconnecting processes.

In order to do that, we need some time until reconnection kicks in.

This should also help with server-initiated connection termination
in general.

References #95.