Make Bunny::Exception inherit from StandardException
Naked rescue like this

begin
  # ...
rescue => e
  # ...
end

catches only descendents of StandardError. Most people don't
know this and this is a very counter-intuitive practice, but
apparently there is code out there that can't be changed that
depends on this behavior.

This is a breaking API change but hopefully won't affect
many.

References #116.