Trying to speed up Bunny::Session#close on JRuby (yes, really)
On JRuby, terminating main reader loop is tricky because sun.nio.ch.KQueueArrayWrapper#kevent0
is a native event that can't be easily intrrupted (without JRuby-specific code
that rips into the guts of RubyThread, anyway).
So it may hang until the kevent call times out.

This does not affect MRI (stress test for connection.close passes) but
not yet complete for JRuby (about 1% of stress tests fail).