Deprecated the `:ack` parameter in favour of `:manual_ack`
* Corrected documentation for both `:ack` and `:manual_ack`
* Added deprecation warning when the `:ack` parameter is used
* Added test to show getting a deprecation warning when using `:ack`

The intentions of this patch are to keep the existing functionality of
being able to pass either `:ack` or `:manual_ack` in as parameters to
`Channel#basic_get` or `Queue#subscribe`. So this should not break any
existing applications that use Bunny. However with this patch, anyone
using `:ack` will encounter a warning saying that the use of `:ack` is
now deprecated.

Signed-off-by: Henry Jenkins <henry.jenkins@aurainfosec.com>