Document Session options for logging.
Document Session#initialize options:

* :logger
* :log_file
* :logfile
* :log_level

One of the guides mentions :logfile and :log_level, but not :logger
nor :logfile.

    http://rubybunny.info/articles/connecting.html

I made the assumption that :logfile is deprecated and documented it as
such.  However, since :logger is useful to me, I decided it must have
been accidentally omitted from the API docs and guide.