Add addresses to connection options
Before this the connection options only allowed multiple hosts, an
address is a combination of a host and a port. This makes it possible to
specify different hosts with different ports.

Allow addresses when initializing the session

An address is a host combined with a port.

Adds some option validations and syntax cleanup

Connection options are now validated to not have both a single and
multiple hosts. Or hosts and addresses.

Removes whitespace from address arrays.

Cleans up addresses_from method

Adds extra spec for incorrect connection options

Adds an additional spec where a single host with a set of addresses is
tested to raise an error.

Refactors addresses_from for readability

The double "host || hostname" wasn't very readable.

Update connection param validations

A single host combined with multiple hosts should log a warning instead
of raising an error. This way it will not break apps currently doing
this.

Combine hosts and addressess in connection options

This way hosts can be in the addresses parameter and vice versa.

Refactors addresses_from method