Make TLS tests pass locally with tls-gen
From now on we ignore spec/tls/* certificate files and assume
they are generated using an environment-specific script
(driving tls-gen, since there is no reason not to).

This also moves a couple of TLS examples to the TLS group which
is excluded by default since not all nodes are TLS-enabled.