Delete a chronically unstable test
I know, I know. Unfortunately it's been a real struggle to get it
to pass in various CI environments, even after simplification,
numerous parameter tweaks and investigations of what kind of concurrency
hazards and time-dependent issues there may be.

This test provides limited value but is the top offender when it
comes to CI build failures that are time consuming to reproduce.

Per discussion with several RabbitMQ team members.