Don't do the closed-then-opened assertion
It is very timing sensitive in the first part. We've
seen this not work in March Hare and other clients.
Yes, it's not perfect but this test is short running
and it only sanity checks connection recovery anyway.

Without this change we get a test or two failing because
it never observes all connections to be closed but debug
logging suggests that everything recovers as expected.