Change recovery attempt limit behavior to match #408
Previous behehavior is not unreasonable but is not what
many users and even RabbitMQ team members come to expect.
Therefore it can be treated as a bug.

Previously a reconnection counter was preserved between
successful recoveries. This made the integration test
that uses server-sent connection.close possible.

With this change, the counter is reset after successful
reconnection but there's an option to go back to the original
behavior. We also do a hell of a lot more logging.
Note that connection.close recovery with a running node will
no longer work for testing this functionality. There aren't
may ways to stop and start the node that aren't
environment-specific, perhaps one day we will mess
with nodes in Bunny tests and introduce a replacement test.

References fff7f4465d2261e8bfbb840d9012c8a13edb9929.