Breaking public API change: Make Bunny::Channel#close raise an exception on a closed channel
The current behaviour is to time out on a network request, which makes no sense
at all, as #528 explains. This is also what other clients do. Making the method
a no-op was also an option but doing so might make genuine channel-level exceptions
harder to discover for the user.

While at it, report the last seen channel.close exception. This is both
informative and consistent with what RabbitMQ Java client does: it
provides a "shutdown reason" (e.g. a protocol exception if any) together
with the "already closed" exception.

Since this is a minor breaking public API change, it won't be backported
to 2.7.x.

Fixes #528.