Remove tls_ca_certificates variable
`tls_ca_certificates` is only used once and it may be that it is not
even used depending on the `if` condition. Instead of assign it, we can
just call the method when it is needed.

(cherry picked from commit 0a415989c01f3b5b661a311f1dc05ff29e3ea509)