Fixnum isn't defined anymore in Ruby version 3.2
Numbers that can not be represented as a 32bit network number will be
silently corrupted. This was already true for Fixnum instances, only
instances of Bignum raised an error. So the code is slightly more
incorrect than before. But I suspect not a lot of people actually tried
sending a Bignum in a header, so I think it's OK.