Notify on connection recovery fail
It’s possible to specify the recovery_attempts option.
Hitting the recovery attempts limit without a successful recovery
leads to a cleanup of connection, etc.
However, the client is not informed about this state and has no chance
to react upon it.

Here we introduce a callback which is called in exactly the above
described situation enabling client code to react to a failed recovery
process.