Introduce a Bunny::Channel setting for cancelling of all consumers before closing
Closing a channel cancels all consumers on it and requeues
all outstanding deliveries for consumers that used
the manual acknowledgement mode.

However, some prefer to avoid having to deal with
the potential in-flight deliveries right before the
connection is closed.

Now Bunny has such an option and every projects
that builds on top (Sneakers, Hutch, and numerous other ones)
wouldn't have to reinvent this wheel.